refactor: move truncate segmentation values
