+++ sdk/src/main/java/ly/count/android/sdk/ModuleEvents.java
@@ -440,7 +440,7 @@ public class ModuleEvents extends ModuleBase implements EventProvider {
 
                 L.i("[Events] Calling recordEvent: [" + key + "]");
 
-                Utils.truncateSegmentationValues(segmentation, _cly.config_.sdkInternalLimits.maxSegmentationValues, "[Events] recordEvent,", L);
+                UtilsInternalLimits.truncateSegmentationValues(segmentation, _cly.config_.sdkInternalLimits.maxSegmentationValues, "[Events] recordEvent,", L);
 
                 eventProvider.recordEventInternal(key, segmentation, count, sum, dur, null, null);
             }
