+++ sdk/src/main/java/ly/count/android/sdk/Utils.java
@@ -13,7 +13,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.security.SecureRandom;
-import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -310,33 +309,6 @@ public class Utils {
             return true;
         } else {
             return false;
-        }
-    }
-
-    /**
-     * Checks and transforms the provided Object if it does not
-     * comply with the key count limit.
-     *
-     * @param maxCount Int @NonNull - max number of keys allowed
-     * @param L ModuleLog @NonNull - Logger function
-     * @param messagePrefix String @NonNull - name of the module this function was called
-     * @param segmentation Map<String, Object> @Nullable- segmentation that will be checked
-     */
-    static void truncateSegmentationValues(@Nullable final Map<String, Object> segmentation, final int maxCount, @NonNull final String messagePrefix, final @NonNull ModuleLog L) {
-        if (segmentation == null) {
-            return;
-        }
-
-        Iterator<Map.Entry<String, Object>> iterator = segmentation.entrySet().iterator();
-        while (iterator.hasNext()) {
-            if (segmentation.size() > maxCount) {
-                Map.Entry<String, Object> value = iterator.next();
-                String key = value.getKey();
-                L.w(messagePrefix + ", Value exceeded the maximum segmentation count key:[" + key + "]");
-                iterator.remove();
-            } else {
-                break;
-            }
         }
     }
